+++ docs/CHANGE_LOG.md
@@ -5,12 +5,14 @@
 #### **New Features:**
 - Added specific properties watch support! This will only watch for changes on specific properties on an object, instead of all of them by default.
 
+#### **Binding Options / Function Options:**
+- Added a new binding/option called "propertyNames", which states the property names that should be watched for changes (defaults to all).
+
 #### **Binding Options / Function Options - Custom Triggers:**
 - Added a new binding/option custom trigger called "onRemove", which states an event that should be triggered when a DOM element is removed.
 
 #### **Public Functions:**
 - Added new public function "resumeWatch()", which is used to resume a watch that has been paused.
-- Added new parameter "propertyNames" to the "watch()" public function, which states the property names to watch changes for (defaults to all).
 
 #### **Fixes:**
 - Fixed a fault that allowed comparisons between arrays and objects (was causing some script errors).